Convert CollapsibleSection to TS
