--- src/CollapsibleSection/index.js
@@ -1,2 +1,0 @@
-import CollapsibleSection from './CollapsibleSection';
-export default CollapsibleSection;
